må runde av grad før utbetalingslinjer bygges
fordi spenn har konvertert grad til int uten avrundning  må eksisterende utbetalingslinjer rundes ned,
fordi vi må reflektere den verdien som OS har fått.

+++ sykepenger-model/src/main/kotlin/no/nav/helse/person/PersonObserver.kt
@@ -62,7 +62,7 @@ interface PersonObserver {
                 val tom: LocalDate,
                 val sats: Int,
                 val beløp: Int,
-                val grad: Double,
+                val grad: Int,
                 val sykedager: Int
             )
         }
