+++ sykepenger-model/src/main/kotlin/no/nav/helse/person/UtbetaltEventBuilder.kt
@@ -147,7 +147,7 @@ private class UtbetaltEventBuilder(
         satstype: Satstype,
         beløp: Int?,
         aktuellDagsinntekt: Int?,
-        grad: Double?,
+        grad: Int?,
         delytelseId: Int,
         refDelytelseId: Int?,
         refFagsystemId: String?,
